Added relative date handling
closes: https://github.com/NexesJS/NQL-Lang/issues/9

This adds relative date expressions in the form now-1d.
These are translated into absolute dates inside the NQL parser
The output format then contains a standard gt/lt type date expression

This is because mongo queries don't support relative dates without using $expr and knex doesn't support them at all.
We'd have to change a lot of stuff to support relative queries to the DB / to JSON but this solves the immediate problem simply.

This has meant introducing state into the lexer, which is new to me. This done so that we can correctly interpret the -/+ symbols
in two different "states". This could be used to improve the lexer further.

This is quite a complex addition to the language, but it reduces complexity everywhere else.